nested the .header-logo and .header-nav in the .header-bar div.
displayed the content to flex and justified-content to stretch.
removed the text decorations and color on anchor and list styles to none.

nested the .header-content and .header-image in the ,header-content div.
displayed the content to flex and justified-content to stretch.
made a width of 50% to the each of the .header-content.
made the .header-image to have a reasonable width and height.
edited the .header-image content (the image goes here) to text-align to center and the its margin to 110, 50, 100, 50 pixels.
changed the sign up input into a button and removed the border and background color changed to blue.